v 0.0.4 WIP
Cleaned up project setup and sequential build of projects.

Update Core base classes to be compatible with CUDA.
Added core class zDate for date utility methods.

Added Cuda toolset projects , for classes using cuda functionality.
Added solar analysis toolset class. under development.

+++ cpp/headers/zCore/base/zDefinitions.h
@@ -286,7 +286,7 @@ namespace zSpace
 	*/
 	struct zEPWData
 	{
-		double dbTemperature, humidity, windSpeed, windDirection, radiation, pressure;
+		float dbTemperature, humidity, windSpeed, windDirection, radiation, pressure;
 	};
 
 	/*! \struct zLocation
